Separate firebase stuff into different directory
