+++ src/firebase/config.js
@@ -1,0 +1,11 @@
+var firebaseConfig = {
+	apiKey: "AIzaSyDqk-tCkJUNKqN1rwhHDvOUrMQ80KDdnYI",
+	authDomain: "fouronesixnine-notes-app.firebaseapp.com",
+	databaseURL: "https://fouronesixnine-notes-app.firebaseio.com",
+	projectId: "fouronesixnine-notes-app",
+	storageBucket: "",
+	messagingSenderId: "967381980154",
+	appId: "1:967381980154:web:5a5a8276ddab77a1"
+};
+
+export default firebaseConfig;
